Take Hetzner DNS + letsencrypt as far as we can
When we try to enable letsencrypt for an app, the acme server is unable
to resolve Hetzner DNS propagation information. Hetzner's DNS services
appear to be unreliable. The move might be to go back to Cloudflare
and see if letsencrypt can support multi-level subdomains even if
Cloudflare's own SSL cert doesn't.